sort_model_packages * com.mmestre -> com.mmestre.rijksmuseum: added intermediate folder to avoid repetition, in the future, of the packages. * com.mestre.model: the classes inside the package were moved to be split in different packages inside, depending on which API method uses that. * All the classes affected by the package move.
