--- src/main/java/com/mmestre/model/Acquisition.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Acquisition {
-    private String method;
-    private String date;
-    private String creditLine;
-}
