--- src/main/java/com/mmestre/model/AdLibOverrides.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class AdLibOverrides {
-    private String titel;
-    private String maker;
-    private String etiketText;
-}
