--- src/main/java/com/mmestre/model/CollectionDetailsResponse.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class CollectionDetailsResponse {
-    private long elapsedMilliseconds;
-    private ArtObject artObject;
-    private ArtObjectPage artObjectPage;
-}
