--- src/main/java/com/mmestre/model/CollectionImageResponse.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-import java.util.List;
-
-@lombok.Getter @lombok.Setter
-public class CollectionImageResponse {
-    private List<Level> levels;
-}
