--- src/main/java/com/mmestre/model/Colors.java
@@ -1,7 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Colors {
-    private int percentage;
-    private String hex;
-}
