--- src/main/java/com/mmestre/model/ColorsWithNormalization.java
@@ -1,7 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class ColorsWithNormalization {
-    private String originalHex;
-    private String normalizedHex;
-}
