--- src/main/java/com/mmestre/model/Dating.java
@@ -1,10 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Dating {
-    private String presentingDate;
-    private long sortingDate;
-    private long period;
-    private long yearEarly;
-    private long yearLate;
-}
