--- src/main/java/com/mmestre/model/Dimension.java
@@ -1,10 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Dimension {
-    private String unit;
-    private String type;
-    private String precision;
-    private String part;
-    private String value;
-}
