--- src/main/java/com/mmestre/model/Label.java
@@ -1,10 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Label {
-    private String title;
-    private String markerLine;
-    private String description;
-    private String notes;
-    private String date;
-}
