--- src/main/java/com/mmestre/model/Links.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Links {
-    private String self;
-    private String web;
-    private String search;
-}
