--- src/main/java/com/mmestre/model/Tile.java
@@ -1,8 +1,0 @@
-package com.mmestre.model;
-
-@lombok.Getter @lombok.Setter
-public class Tile {
-    private long x;
-    private long y;
-    private String url;
-}
