+++ src/main/java/com/mmestre/rijksmuseum/model/common/Acquisition.java
@@ -1,0 +1,8 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Acquisition {
+    private String method;
+    private String date;
+    private String creditLine;
+}
