+++ src/main/java/com/mmestre/rijksmuseum/model/common/Colors.java
@@ -1,0 +1,7 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Colors {
+    private int percentage;
+    private String hex;
+}
