+++ src/main/java/com/mmestre/rijksmuseum/model/common/ColorsWithNormalization.java
@@ -1,0 +1,7 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class ColorsWithNormalization {
+    private String originalHex;
+    private String normalizedHex;
+}
