+++ src/main/java/com/mmestre/rijksmuseum/model/common/Dating.java
@@ -1,0 +1,10 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Dating {
+    private String presentingDate;
+    private long sortingDate;
+    private long period;
+    private long yearEarly;
+    private long yearLate;
+}
