+++ src/main/java/com/mmestre/rijksmuseum/model/common/Dimension.java
@@ -1,0 +1,10 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Dimension {
+    private String unit;
+    private String type;
+    private String precision;
+    private String part;
+    private String value;
+}
