+++ src/main/java/com/mmestre/rijksmuseum/model/common/Label.java
@@ -1,0 +1,10 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Label {
+    private String title;
+    private String markerLine;
+    private String description;
+    private String notes;
+    private String date;
+}
