+++ src/main/java/com/mmestre/rijksmuseum/model/common/Links.java
@@ -1,0 +1,8 @@
+package com.mmestre.rijksmuseum.model.common;
+
+@lombok.Getter @lombok.Setter
+public class Links {
+    private String self;
+    private String web;
+    private String search;
+}
