+++ src/main/java/com/mmestre/rijksmuseum/model/details/AdLibOverrides.java
@@ -1,0 +1,8 @@
+package com.mmestre.rijksmuseum.model.details;
+
+@lombok.Getter @lombok.Setter
+public class AdLibOverrides {
+    private String titel;
+    private String maker;
+    private String etiketText;
+}
