+++ src/main/java/com/mmestre/rijksmuseum/model/details/CollectionDetailsResponse.java
@@ -1,0 +1,10 @@
+package com.mmestre.rijksmuseum.model.details;
+
+import com.mmestre.rijksmuseum.model.common.ArtObject;
+
+@lombok.Getter @lombok.Setter
+public class CollectionDetailsResponse {
+    private long elapsedMilliseconds;
+    private ArtObject artObject;
+    private ArtObjectPage artObjectPage;
+}
