+++ src/main/java/com/mmestre/rijksmuseum/model/image/CollectionImageResponse.java
@@ -1,0 +1,8 @@
+package com.mmestre.rijksmuseum.model.image;
+
+import java.util.List;
+
+@lombok.Getter @lombok.Setter
+public class CollectionImageResponse {
+    private List<Level> levels;
+}
