+++ src/main/java/com/mmestre/rijksmuseum/model/image/Level.java
@@ -1,0 +1,11 @@
+package com.mmestre.rijksmuseum.model.image;
+
+import java.util.List;
+
+@lombok.Getter @lombok.Setter
+public class Level {
+    private String name;
+    private long width;
+    private long height;
+    private List<Tile> tiles;
+}
