+++ src/main/java/com/mmestre/rijksmuseum/model/image/Tile.java
@@ -1,0 +1,8 @@
+package com.mmestre.rijksmuseum.model.image;
+
+@lombok.Getter @lombok.Setter
+public class Tile {
+    private long x;
+    private long y;
+    private String url;
+}
