+++ src/test/java/com/mmestre/rijksmuseum/RunCucumberTest.java
@@ -1,0 +1,17 @@
+package com.mmestre.rijksmuseum;
+
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectPackages;
+import org.junit.platform.suite.api.Suite;
+
+import static io.cucumber.junit.platform.engine.Constants.FEATURES_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectPackages("com.mmestre.rijksmuseum")
+@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
+@ConfigurationParameter(key = FEATURES_PROPERTY_NAME, value = "src/test/features")
+public class RunCucumberTest {
+}
